Improved performance of Paritlce updates by about 10%
In this commit, I focused everything on performance. To do this, I began by calculating the number of floating operations performed in both the Update_P and Update_x functions (per particle). To do this, I counted the number of floating point operations in each vector/tensor method, as well as the number of times each operation was used. This revealed that update_P uses about 10,000 FLOP while update_x uses around 30,000 FLOP. This explains why the update_P and update_x functions take up about 25% and 75% of the runtime respectivly.

I tried optimizing a few of the operations that were used most often, but to no avail. It seems that my tensor/vector methods are running as quickly as they can (maximum efficiency). Thus, to improved performnace, I need to reduce the total number of operations.

I began searching for areas where operations can be eliminated. Each particle, on average, has 200 neighbors. Thus, operations in the neighbor loops are performed hundreds of times. Therefore, I placed emaphasis on reducing the number of operations inside of the neighbor loops.

It quickly occured to me that it does NOT make sense to calculate mag_Rj on each iteration. This quantity is lagrangian (and it uses sqrt, whcih takes many operations). Thus, I made a Mag_R array that is dynamically allocated to store the magnitude of R for each neighbor particle. This change took about ~1000 operations (~3%). Next, I noticed that the Hg force was always multiplied by -(.5)*E*Vi*Alpha. This added 4 operations to each iteration of the neighbor loop (nearly 800 operations per particle). I moved this part of the multiplication outside of the loop (see Particle.c:415) , reducing the number of operations by an additional 800 iterations..

I also noticed that the way I calculate delta_ij and delta_ji could be improved. Lines 360-400 explain how I can improve the calculation of these quantities. Making this change removed about 2000 operations. In total, almost 4000 operations (nearly 11% of the total for update_x) were removed. In theory, this should make the update_x function run about 10% faster. I tested with the changes, and observed roughly a 10% increase in performance (the update_x runtime before making the change was 309.8 seconds, runtime after making the change was 273.8 seconds, an 11.6% improvement!)

I also redid the timing tests. I focused on the operations that are used the most in the update_x and update_P methods. I wrote the new tests so that they could be run on an O3 compiler optimization. This fundamentlaly changed the way that these tests ran, so the old tests were removed.

Finally, I now print E rather than e (E = (1/2)(C-I)) to the output files.