Added a specialized 'Box' neighbor generating function, Can now run general cuboid tests
I ran the 512k particle test overnight last friday. Impressivly, the actual Particle iterations only took about 3.5 hours (using a smaller support radius apparently really improves performance). However, the program spent more than an hour generating the neighbor list. this is because each particle searches through each other particle when generating its neighbor list. This means that for a N particle test, there are N*N checks to make the neighbor lists. Thus, the runtime of the neighbor list generating function grows quadratically.

However, if we're running tests with boxes/cubes, then we should be able to take advantage of the geometry to generate our lists. I made a new function 'Generate_Neighbor_Lists_Box' that works just like the old 'Generate_Neighbor_Lists' function except that it is designed for cuboid particle arrangements. Rather than searching the entire array, the x, y and z dimensions of the array are passed to the function along with the support radius (in units of inter particle spacings). The function then only checks the particles in a small cube about the current particles. If the support radius is 8, then we check the 9x9x9 cube with the current particle at the center. This gives the neighbor list function a linear runtime relationship.

I also redid the Particle tests to work with a general cuboid rather than a cube. This was done by adding side Lengths for the x, y and z dimensions. It should be noted that there are y_Side_Len*z_Side_Len particles in a layer and z_Side_Len particles in a column. With these changes (and a number of bug fixes related to improper use of the particles array) I got the new code running.

All the changes can be seen in the Tests.c file.

I also added a 'SUPPORT_RADIUS' definition this reports the support radius in units of inter particle spacings . This definition is located in the SPH_Main.cc file and is used to generate particle lists + set up the particles h static parameter.