Switched from z-y-x major order to y-z-x major order
I believe that it makes more sense to store the particle in 'vertical column' major order. This basically means that the (1,1,1) and (1,2,1) particle are next to one another in memory. In the old system, the (1,1,1) and (1,1,2) particles would be next to one another in memory.

So why make the change? I choose to make this change because I believe that the vertical direction is the most important. This is where gravity acts, among other things. Thus, when printing out data to files, I am much more interested in looking at how a property varies about a vertical column than in a horizontal row (fixed x, y). This controls how particle data is prointed to the debug files/makes it easier to understnad what's going on.