Singular tensor returns input tensor, renamed particle debugger
I renamed the particle debugger file to 'Particle_Debugger' (removed the 'file')

I also made the 'tensor is singular ' error return the invoking tensor rather than some random other tensor (see tensor.c:649).

I tried running my code with a larger set of particles and a few of them blew up. The corners in particular broke for no good reason. I need to figure out why this happened and come up with a way of fixing it. Also, in the current test the slice was way too thin. Multiple layers need to be removed for this to work properly.