Improved Contact
I continued to develop the contact force. I added a description to explain how the contact algorithm actually works.  I also played around with the k constant (this actually depends a LOT on the density of the particles... make the density too high and the particles fall right through the barrier).

I also tried optimizing the contact algorythm. To begin, I noticed that it every one of Body_B's x vectors is accessed Num_Particles_A times. However, since the particles are much larter than a cache line, the computer has to pull in a new cache line each time it changes i or j (even though x only takes up 3 double precision numbers). To fix this, I made an array 'Body_B_x' which stores only the x vectors of the Body_B array. This array is then used to access x for body B rather than Body_B[i].x. This method improved the runtime of the Contact kernel by about 10% (for a 8000 particle sample).

I also tried grouping sacalars together to minimnize the number of vector operations (ee Particle_Contact.cc:43). These changes slightly improved runtime.

I also removed the IPS argument from the Contact function (this argument was not being used).

 I also found that removing alias variables for double members (in Upate_x and update_P) slightly improved runtime (by roughly 3%). I'm not totally sure if this change was just noise or actually implies some improvements. Regardless, the change improved runtime. (See Particle_Update.cc).

I also quadrupled the size of the barrier (see Tests.cc). This required making a few changes to how the Boundary array is populated. I also found an error in my test code: I was not multiplying the boundary X vector by IPS. I believe this error was actually causing the asymmetry observed last week. To understand why, let IPS = .5 and let the side length be 10. Without multiplying X by IPS, the boundary extends over {x,z} = {(0,9), (0,9)} while the bottom face of the cube only extends over {x,z} = {(0,4.5) , (0,4.5) }. Because of this, the boundary extends beyond the cube on two sides (the +x and +z side) but not on the other two (the -x and -z side). When the cube hits the boundary, it flattens out. Without The extra support in the +x and +Z directions ensures that the part of the cube that flattens out in these directions is supported by the boundary. However, the parts of the cube that flatten out in the -x and -z directions fall over the edge of the barrier. Therefore, the cube is more supported in the +x and +z directions than in the -x and -z directions. This causes the cube to lean twoards the -x and -z directions, causing it to fall off the boundary after a few bounces. With the changes that I made today, these issues are no longer present.