Optimized particle parameters for zero-viscosity
I attempted to stabilize my code with zero viscosity. To do this, I varied three variables: Lame, Mu0, and alpha. E was also changed, but was always set to the equivalent young's modulus for the given Mu0/Lame values (so it was not an independent variable).

Right off the bat, I found that the Lame paramater was way too large, and that Mu0 was way too small. I found that my code was far more stable with a lame value around 1. After making this change, I systematically attempted to optimzie each of Lame, Mu0 and Alpha. The results from this effort are listed in a new spreadsheet titled 'Improving zero viscosity stability'. I used two metrics to measure how  'stable' my code was: Number of steps before divergence and number of steps before cube is unrecognizable. The second metric is subjective (where do we draw the line between what looks like a cube and what does not) while the first is objective. Using these metrics, I found that the following values basically give the most stable code: Lame = 1.125, Mu0 = .275, alpha = 7.5. The stability is very responsive to changes in Mu0 and alpha, but less responsive to changes in Lame. The general rule with Lame is to not make it larger than about 1.25. The stability varies a lot as the value of mu0 is decreased below .75, I believe that this is mostly noise. In general, I think that the stability is basically the same for mu0 less than about 1.

I did not run very many trials with alpha. More trials should be run to further optimize alpha. All that I can say is that the code gets very unstable for alpha much larger than 10. a value of 7.5 was choosen because the cube retained its shape for the longest period of time with this value of alpha (of the tests that I ran).

Finally, K was decreased from 250 to 100. I think that K = 100 produces a nicer looking contact while still prevnting particles from falling through a single particle barrier.

The TODO dev plan has also been updated.