Updated Damage description (comment)
I implemented a ray based particle damage method. This method works off of purely geometric arguments. How does this wrok? We essentially treat each particle as a 'sphere' with a pre-determined radius. This radius is the 'radius' paramater in the particle class (see Particle.h).

When a particle becomes damaged, we have to causally remove that particle from the Particle's array. What this means and how we accomplish it is explained in the rather lengthy comment at the start of the Remove_Damaged_Particle function (see Particle_Damage.cc).

I use purely geometric arguments to determine if P_j is in the shadow region of P_In. Again, see Particle_Damage.cc.

The tests file has also been modified to use this new system of damage (See Tests.cc).

As is explained above, I added a 'Radius' member to the Particle's class. This radius simply determines the 'size' of a given partilce, which is used to calculate shadow regions (see Particle.h, Particle_Damage.cc).