Added a 'Remove_Partile' method, fixed ripping
WARNING: this code currently crashes the program soon after ripping begins (need to fix).

I tried making the support radius 4. Doing so caused the program to violently crash. I eventually discovered the issue was that I was making a faulty assumption in my code: I assumed that if particle P_In has particle P_i as a neighbor and P_i has a neighbor P_j that is in the shadow region of P_In, then P_j must ALSO be a neighbor of P_In. This assumption is false. The reason is that a 4th particle can become damaged/eliminate P_In and P_j's connection but not P_In and P_i's neighbor status. The fix to this I added a new function called 'Remove_Neighbor'. The idea is that when P_i finds a neighbor, P_j, that is in P_In's shadow region, we remove P_i from P_j's neighbor list. By doing this, we ensure that the shadow region effect is ALWAYS symmetric. In other words, if A stops being a neighbor with B then B stops being a neighbor with A.

This new remove_neighbor function can be found in the Particle.cc class. I also added a small description to Particle_Damace.cc:214 that explains the basic idea of how the Remove_Neighbor function works, though a better explanation can be found in Particle.cc.