Fixed bug
I found the bug that was present in the previous commit. The issue was that I wasn't zero initializing the New_A tensor. When calculating A I use compound tensor-tensor addition. As a result, in some cases, the new A tensor was NOT zero initialized and was therefore doing terrible things. The fix can be seen in Particle.cc:151.

With this fix, the code now runs with h = 4 with a 10x30x10 column.

I also removed an annoying variable (see Particle_Damage.cc).