Improved breaking test
I increased the height of the column from 30 to 31 layers. This means that the two ends of the break (the two pieces that break away) are equal in length. (15 above and 15 below). This leads to a more symmetric rip.

I also improved the way that the 'cut' is generated. In the old system, I located the 'cut' particles away from the main body. This worked alright, but created a problem. Particles above and below the cut were still neighbors. this means that after the body separated, the particles above and below the cut were still connected and would feel strains. This eventually caused these particles to fail.

To fix this, I insted 'pre-damated' the cut particles. The body is generated normally (not displaying the cut particles). After the neighbors are set up, I damage and then remove the particles along the cut (using Remove_Damaged_Particle). This way, particles above and below the cut are not connected. Making this change fixed the issue described in the paragraph above, making the ripping test run cleaner (fewer broken particles) and more accuretly (no weird connections that shouldn't be there). The changes can be seen in Tests.cc. As a result of tis change, I no longer ignore the cut particles when generating neighbors/updating P/updating x (see Tests.cc)