Reorganized code, built a dedicated 'simulation' function
I realized that the time had come to move the 'Particle test' code into its own dedicated file. This is the new 'Simulation' files (source + header).

At first, I moved the contents of the Particle test into the simulation source file and most of the contetns of the test header into the Simulation header. The old Particle_Test code was placed into a new function 'Run_Simulation'. This function is inside of the new 'Simulation' namespace. This seemed to be working well, but quickly created issues. In particular, it was not possible to make the 'Run_Simulation' function a friend of the particle class and have the Simulation namespace set static particle class members. To fix this, I created a third new file, the 'Namespaces.h' file. This file sort of works like a forward declaration for all the namespaces in my code. Every one of the namespaces in my code is named in this file, and a few namespace function prototypes are there as well. Doing this allows the rest of my code to know which namespaces exist. This fixed my issue.

I also added a few extra print statements to the simulation code (like printing out the particle class static members before th esimulation runs, as well as the cuboid paramaters).

I also separated the 'declaration' and 'assignment' portions of the static particle class members (see Simulation.h). The 'default assignment' has been placed inside of a function called 'Set_Static_Particle_Members'. This function, as the name suggests, sets the particle class static members if the program is generating a new simulation/not reading from an existing one.

I also added a new static member to the particle class, the 'support radius'. This simply stores the kernel function support radius in units of Inter Particle Spactings rather than in mm. This was done to replace the old 'SUPPORT_RADIUS' global variable. I thought this variable should be a particle class member.

I plan on splitting the main 'Run_Simulation' function into multiple sub functions/making the code easier to read.