Added 'getters' to Particle class
I added a nuber of 'getter' methods to the particle class. These methods return the values of a number of private particle variables. I used these methods to remove a number of the particle's freinds (friends are dangerous). I was able to unfriend the 'Find_Neighbor', 'Print_Particle_To_File' and 'Export_Particle_Positions' methods. Most of the remaining friends rely on particle class variables (static members), so removing their friend status is not as easy.