Fixed some units issues
I finally resolved the old "why are all the forces in units of N*mm^3?" question. The answer, as it turns out, was that I had made an error in calculating my units. The shape function, W, has units of 1/mm^3, and the gradient of W has units of 1/mm^4. With this fix, the forces are indeed in units of N. There was no issue after all. I fixed this mistake throughout my code (most of the changes are jsut comments)