Reorganized files
I made two new directors, 'Code' and 'Files'. The code folder holds all of the program's code (and is where the code should be compiled). The 'Files' folder houses all of the input/output files including the 'Particle_Data' file, the 'Force_Files' (in their own directory) and the 'Position_Files' (also in their own directory).

This reogranization was done to make the SPH folder more managable.