Started writing a 'FEB' reader
I started writing a function to read in node data from a FEB type file. This code has NOT been finished/does not work yet.

I made a new namespace for the file reader called FEB_Reader. This  namespace currently only has one function, 'Read_FEB_File', which is designed to do what the name implies. There is an issue, however, since feb files don't specify the node count. Thus, I am using the following stragety: Read in one line at a time. For each line, check if it contains the '<Nodes>' word, implying the start of the nodes section. Once we have found the nodes section, read in the ID's. However, before we can read in a node id, we need to make sure that we're not at the end of a line. Record the position of the start of the line, read a line into file, check if that line contains "<\Nodes>". If it does, end. If it doesn't, then we're on a line that has a node id. Go to the start of the line, and read in the ID using a known format (see FET_File::78). Once we are at the end of the nodes, we go back tothe start of the nodes section. The ID variable now stores the nuber of nodes so we know how many lines to read in.

I need to write the actual read in part.