Added FEBio file reader, updated data dump
I got the FEBio file reader working. To do this, I first modified the 'Read_FEB_File' function to accepet the file name (as a string), a pointer to a pointer to a vector and a reference to the number of nodes. This function first reads through the file until it finds the starts of the nodes section (designated by the phrase <Nodes>). Once this is found, the code moves through the nodes, keeping track of the largest ID number.

To do this, I first write down the start of a given line (using ftell), then read in a line, then check for the end of nodes section keyword (which is signaled by the keyword </Nodes>), then go back to the start of the line, then read in an ID (if not at the end of the nodes section), then go back to the end of the current line. see FEB_File.cc:66-94.

Once we know the number of nodes, the code goes back to the start of the Nodes section. Here is where that pointer to pointer to vector comes in. X_Ptr is the address of the X pointer variable in the calling code. We want this variable to point to an array of vectors (that stores the reference positions of each read in node). Thus, we set the value of the pointed to location (which is the address that X in the calling function points to) to a pointer to the address of a dynamic array of Num_Nodes vectors. We then go through the Nodes section, printing the positions of each node to the new dynamic array. Implementation can be found in FEB_File::100-111.

Using this 'pointer to pointer' method is necessairry to get desired results. Why? Well, in the calling function, we declare a pointer (X), that should point to a dynamic array of vectors. The value stored in this pointer is the address of the start of the dynamic array. Thus, to store that address, we need to  know the address of the pointer itself. Therefore, rather than passing the pointer, we need to pass the address of the pointer. This way, in the FEBio reader function, we assign the address of the dynamic array to the value of X_Ptr. Thus, once the function is finished, X (in the calling function) stores the address of the dynamic array.

I used the same method to update how 'Data_Dump' works (see Data_Dump.cc).