Added Particle_Array class
I added a new class, Particle_Array, to my code. I made a new source and headerfile for this class (Particle_Array.cc and Particle_Array.h, respectivly).

The Particle_Array class, as the name implies, is used to store an array of particles. The static Particle class members have been moved to the Particle_Array class. The idea here is that each Particle_Array object can have its own material parameters (the static particle members). By doing this, we can create multiple bodies with different material parameters.

The new particle_array class will be used to help create the needle.

The partilce array class has Setting and Getting methods (to set and get the old Static Particle class members), as well as a [] operator overload. Each array knows how many particles it has. The member array is dynamically set whenever the particle_array is given a number of particles. It should be noted that there is no set method for the shape function amplitude or h. This is because these quantities can be derived once the support radius and inter particle spacing is known. Thus, to make it easier to work with particle arrays, the user is only allowed to set the IPS and support_radius. h and the shape function amplitude are caclulcated/stored in the particle array. There are 'getters' for both h and the Shape_Function_Amplitude (since some functions are more logically expressed in terms of these units than IPS or Support_Radius).

This new class required substantial changes to my code. All statements that read or wrote the particle class static members have been removed/replaced with the appropiate set/get method (for a particle array class). All arrays of particles have been replaced with Partilce_Array objects.

The main idea behind this chnage is to allow my code to run with multiple bodies, each which have VERY different material and kernel parameters. This was not possible before when the material paramaters (and kernel paramaters) were fundamental properties of the particle class (Static members). It was  not possible, for example, to have one body with a particular young's modulus and another body with a different young's modulus (in the old code). Such a situation is easily acheived using Particle_Arrays (just declare two particle_Arrays, give the two different E values).

This switch actually allowed for some functions to become a lot simplier. For example, the 'Contact' method use to requre 5 arguments (see Particle_Contact.cc). further, the old code did not take allow for the different bodies to have different support radius's (h). when written in terms of Particle_Arrays, however, only two arguments are needed (the two particle arrays). The number of particles and support radii are carried with these objects/can be determined without explicit arguments for them!