Added a 'SetUp_FEB_Body' method
I created a new method to essentially create a needle from a febio file. The code for this can be found in Simulation.cc:374.

Currently, the code is set up to drop a needle into a 20x20x20 blob (note: the needle and blob are made out of the same material).