Started updating Data_Dump code (for multibody)
I am in the process of updating the Data_Dump code for multiple bodies. The idea here is to save files using particle_array's names. Under the new system, the plan is to create a single 'Particle_Data' file. This file will store the number of particle arrays, their names, and the number of particles in each. Each Particle array will be stored in its own file (using the same format as the current particle_data file).

the read and write functions work differently with this system. The read files will need to accept an array of particle_arrays as well as the number of elements in that array. This information (along with each array's name) will be used to write the new particle_data file (with number of arrays, names of each array, and number of particles in each array). Each partilce_array's file will be named using the particlE_array's name.

The read/load functions will accept a pointer to a pointer of particle arrays (the pointed to array will dynaically be allocated into the array of particle arrays/bodies). The program then reads in the particle_data file, which tells the program the number of particle arrays as well as their names/sizes. We use this information to then read in the indivudual particle_array files/load them using the old 'Load_Particle_Array_From_File' function.

The code to implement this  new standard is only partially written (I will finish it tomorrow)