Updated Print_Particle_Forces function
I updated the Print_Particle_Forces function. The new version copies what I did for the Print_Particle_Positions file. Basically, the save file is created using the name of a Particle_Array. This allows me to print force files for multiple bodies in a single simulation. To make this work, I first read in the Particle_Array's name. This name is then compared to each element in a linked list. If the name does not match any name in the list, then the name is added to the end of the list. I also add an element to a second list of unsigned integers. This second list keeps track of how many times we've printed force information on this particular body. If a particle_array's name is already in the name list, then the corresonding element of the file counter list is incremented and a new file is created. The code appears to run.