Updated the 'Setup_Body' methods (in simulation.h).
I eliminated the Setup_Body and Setup_Boundary methods. I have replaced them with a single method: 'Setup_cuboid'. I realize that my code can only set up a Particle array if that array is a cuboid. I used this knowledge to redo some of my simulation code. For one, cuboid bodies (boundary or not) call the setup_cuboid method. further, I realized that Partice_Arrays that are not cuboids or from file must be boundaries (because I cant set them up otherwise). I used this knowledge to add another check to my simulation code (see Simulation.cc:78). I also caught a logical bug in my simulation.cc code. The code was setting up FEB_file bodies as if they were cobuoids (which is very wrong) I fixed this bug (see simulation.cc:100).

I also slightly updated my syntax for functions. I think that functions are much easier to read if there are 3 lines between them (this makes it very clear where 1 function ends and another begins).

The code is now fully revamped (in terms of particle_array objects rather than an array of particles). The code still causes the needle array to blow up... I need to work on that. However, the code is at least as good (or bad) as it was before the update but is now (hopefully) easily modified to work with n continuum bodies.