Added a 'damageable' parameter to Particle_Arrays, added F to printout
The new damageable member of the Particle_Array class, as the name would imply, controls wheater or not a PArticle_Array can take damage. This is used to bypass the damage calculations in Update_P. A new setter/getter method has been added for this new member (In Particle_Array.h), and a new setup field has been added to Simulation.h (called Is_Damageable... see Simulation.h).

Finally, I reliazed that my 'data dump' was not complete. To properly save the system, the deformation gradient F must also be saved. I realized this when watching an early needle simulation. I ran for 40,000 steps, saved, then ran for another 40,000 steps from file (loaded results of first 40k). In the first few steps of the second half of the test a bunch of stressed patticles rushed inwards. This was because when these particles were loaded, their F member (the old deformation gradient) was set to the identity tensor. These particles were highly deforme,d however, so the actual deformation gradient was likely very different from the identity tensor. Therefore, when calculating the velocity, the new F (with deformation and all) looked VERY different from the identity tensor, whcih suggested that the deformation gradient had rapidly changed. This caused a large viscosity to appear, causing the particles to rush in. To fix this, I simply added the deformation gradient to the data dump (see Data_Dump.cc).

I also added the 'Is Damageable' member to the data_dump (for paritlce arrays)... also see Data_Dump.cc.