Added 'Anti-divergence' checks + Shortened needle
I tried figuring out why my code blows up... I find it particulaly interesting that when it diverges everything goes to nan, not infinity. This tells me that something mathematically illegal is going on. Looking throuhg my code, I found two things that could cause a nan result: ln(J) and sqrt(LamM). I believe that the sqrt one is not causing my issues since that value would remain isolated/wouldn't impact x, V, etc... This tells me that something bad is happening with ln(J). I think that it's possible that J becomes negative, in which case ln(J) is indeed not a real number. I therefore threw in a check to detect if J < 0. If it is, then that particle is damaged/remved. I also added a check to see if acceleration is nan. If it is, I damage and remove the particle. I am hoping that the acceleration check never gets used/ that ln(J) is where all my issues are coming from. The changes can be found in Particle_Update.cc.

I had to remove the 'const' qualifier from the Particles argument of update_x (Particle_Update.cc). I also made this function a freidn fo the vector class, which required me to put a function prototype in the Namespaces file.

Finally, I updated the materail model for steel to use a density of 8, not 1 (as it should be... see Materials.h).