Updated viscosity calculation
To calculate the viscosity, we need to calculate F' (time derivative of the deformation gradient). In our old code, we used a simple backwards difference calculation to do this. The issue is, however, that backward difference has O(h) error. Our time integration has O(h^2) error. Thus, our error was being dominated by the viscosity calculation, not the time integration. To fix this, I wanted to give the viscosity calculation O(h^2) accuracy. Doing this required switching to a 3-point derivative approximation. I used the 3 point backwards difference apporixation.

Implementing the 3 point derivative expression, however, created some problems. For one, I needed each particle to store both it's current and past deformation gradient. This meant replacing the 'F' member in the particle class with a 2 element array of tensors called 'F' (see Particle.h). With two deformation gradients, however, I had to rewrite a bunch of code. For one, the 'Get_F' method now accepts an integer argument (i) and then returns the specified F tensor (see Particle.h). Second,  I needed a way of deciding which tensor was 'old' and which was 'new'. There are a number of ways to do this, I did not choose a standard one. Naturally, we may want to make the '0' element of the F array the 'newest' deformation gradient and the '1' element the old one. The implementation here is simple, but the performance is awful. To see why, consider how we 'update' F in this case. first, the '0' element of F is moved to the '1' element, then the new F is written to the '0' element. This requires writing two tensors (slowwwww). Another option is to alternate which element we store the newest tensor in (store new F in the 0 element on even steps and old F in 1 on the odd steps). This requires only one write operation. The most natural way to do this is to keep track of a counter. After each step, we increment the counter. If the counter is even then element 0 has the newest F, otherwise element 1 does. We could use % to determine if the counter is evne or odd but division/modulus is SLOWWWWW. To avoid this, we  use a simple if statement. When incrementing, if the counter is 0, make it 1. Otherwise make it 0. This requires 0 divisions and thereofre runs quickly. I stored the counter in the particle array class (so each body has it's own counter, every particle in that body treats the same F element as the 'new' one). To see these changes, see 'Particle_Array.h').

Finally, with these changes, I had to rewrite some code to work with the array of F's rather than a single F tensor (see Particle_Update.cc, for example). I also added code to increment the counters (see simulation.cc). I also had to change the way that saving works. To do this, I made each particle saves both of its F's, and each particle_Array saves the last value of its counter (see Data_Dump.cc).

With these changes, the code ran nicely. I tried decreasing the viscosity and running the tumble test. The new code is both stable and (with small viscosity) seems to reduce tumbling!