Got parallel code running!
I parallelized my code. This produced a few issues, but I was eventually able to work out the bugs (I think all of them) and get the code running in parallel (with openmp).

To begin, I first analyzed what could be parallelized. I decided to parallelize the timestep loop. Thus, I used a single parallel region around this loop (see simulation.h). I choose to use default shared (this may need to be changed). I tuen parallelzied most of the major kernels inside of the timestep loop.

All of the timer updates were done suing a single construct (with a nowait clause) since only one thread needs to update each timer (otherwise we get a race condition!).

I also parallelized the boundary conditions using a single construct.

I used a parallel for loop to update P. this loop was placed on the 'paralticles loop' (like 262) . Placing the for loop around the 'bodies loop' (like 254) wouldn't help since the number of bodies may be less than the number of threads + the size of each bodies vary greatly. Parallelizing the particles loop helps balance the load (each particle takes the same number of operations to update).

The contact loop (which was the trickies to parallize) was done using a orphaned for loop.  I initially tried parallizing the 'generate_x array' parts of the contact code. This creates an issue, however. Since the contact code is called in parallel, all of the local variables are private. This means that each thread has a private copy of the x array. Thus, to  make the code run properly, each thread must go through every element of body B and copy the x vectors over. I parallelized the particles loop of the contact code. There is, however, a small issue here. It is possible for two particles in A to apply a force to the same particle in B. If this happens in parallel at the same time, then the outcome is non-determinstic. Thus, we must update B's vectors one at a time (thus we use a critical region). This unfortunetly makes the code run quite slow. Currently, the contact kernel is taking up roughly 1/3rd of the runtime, more than even update x. This is especially troubbling since most of the runtime for contact happens in a short period of time. The contact code needs substiantial optimization.

Update x has been parallelized with a parallel for loop (just like update P). I also placed the 'update_F_Counter' in a single construct (for obvious reasons... bad data races!).

The 'write to file' part o the code has been parallelized using a parallel for loop (we can write multiple files at once as long as just one thread works on each file).

The update time_step_counter loop has beeen parallelized using a single construct.

I also redid the way that timing works. I switched to using omp_get_w_time when running in parallel (this is accomplished with some define statements and rearranging of variable definitions.... see Simulation.cc:21-26 + Simulation.cc:358-418 + SPH_Main.cc:
10-19).

With all of these changes, the code ran in parallel. Right now, the code runs about 2x as quick as it does without parallelization. There is plenty of room for optimization.

DAMAGE DOES NOT WORK. I realized there is a serious bug with damage (removing a damaged particle must happen sequentially... my code needs to be modifieid).